% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb-get-adm.R
\name{gb_get_adm}
\alias{gb_get_adm}
\alias{gb_get_adm0}
\alias{gb_get_adm1}
\alias{gb_get_adm2}
\alias{gb_get_adm3}
\alias{gb_get_adm4}
\alias{gb_get_adm5}
\title{Get country files from geoBoundaries for a given administration level}
\source{
geoboundaries API Service \url{https://www.geoboundaries.org/api.html}.
}
\usage{
gb_get_adm0(
  country,
  simplified = FALSE,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  metadata = FALSE,
  quiet = TRUE,
  overwrite = FALSE,
  cache_dir = NULL
)

gb_get_adm1(
  country,
  simplified = FALSE,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  metadata = FALSE,
  quiet = TRUE,
  overwrite = FALSE,
  cache_dir = NULL
)

gb_get_adm2(
  country,
  simplified = FALSE,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  metadata = FALSE,
  quiet = TRUE,
  overwrite = FALSE,
  cache_dir = NULL
)

gb_get_adm3(
  country,
  simplified = FALSE,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  metadata = FALSE,
  quiet = TRUE,
  overwrite = FALSE,
  cache_dir = NULL
)

gb_get_adm4(
  country,
  simplified = FALSE,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  metadata = FALSE,
  quiet = TRUE,
  overwrite = FALSE,
  cache_dir = NULL
)

gb_get_adm5(
  country,
  simplified = FALSE,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  metadata = FALSE,
  quiet = TRUE,
  overwrite = FALSE,
  cache_dir = NULL
)
}
\arguments{
\item{country}{A character vector of country codes. It could be either
\code{"ALL"} (that would return the data for all countries), a vector of country
names or ISO3 country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{simplified}{Logical. Return the simplified boundary or not.}

\item{release_type}{One of \code{"gbOpen"}, \code{"gbHumanitarian"},
\code{"gbAuthoritative"}. Source of the spatial data. See \strong{Details}.}

\item{metadata}{Should the result be the metadata of the boundary?}

\item{quiet}{Logical, on \code{FALSE} it displays information of the call. Useful
for debugging, default is no messages \code{quiet = TRUE}.}

\item{overwrite}{A logical whether to update cache. Default is \code{FALSE}.
When set to \code{TRUE} it would force a fresh download of the source
\code{.geojson} file.}

\item{cache_dir}{A path to a cache directory. If not set (the default
\code{NULL}), the data would be stored in the default cache directory (see
\code{\link[=gb_set_cache_dir]{gb_set_cache_dir()}}). If no cache directory has been set, files would be
stored in the temporary directory (see \code{base::tempdir()}).}
}
\value{
\itemize{
\item With \code{metadata = FALSE}: A \code{\link[sf:sf]{sf}} object.
\item With \code{metadata = TRUE}: A tibble.
}
}
\description{
\href{https://www.geoboundaries.org/index.html#usage}{Attribution} is required
for all uses of this dataset.

These functions are wrappers of \code{\link[=gb_get]{gb_get()}} for extracting any
given administration level:
\itemize{
\item \code{gb_get_adm0()} returns the country boundary.
\item \code{gb_get_adm1()} returns first-level administration
boundaries (e.g. States in the United States).
\item \code{gb_get_adm2()} returns second-level administration
boundaries (e.g. Counties in the United States).
\item \code{gb_get_adm3()} returns third-level administration
boundaries (e.g. towns or cities in some countries).
\item \code{gb_get_adm4()} returns fourth-level administration
boundaries.
\item \code{gb_get_adm5()} returns fifth-level administration
boundaries.
Note that not all countries have the same number of levels.
}
}
\details{
See \strong{Details} in \code{\link[=gb_get]{gb_get()}}.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}

\donttest{
lev2 <- gb_get_adm2(
  c("Italia", "Suiza", "Austria"),
  simplified = TRUE
)


library(ggplot2)

ggplot(lev2) +
  geom_sf(aes(fill = shapeGroup)) +
  labs(
    title = "Second-level administration",
    subtitle = "Selected countries",
    caption = "Source: www.geoboundaries.org"
  )
}
\dontshow{\}) # examplesIf}
}
\references{
Runfola, D. et al. (2020) geoBoundaries: A global database of political
administrative boundaries. \emph{PLOS ONE} 15(4): e0231866.
\doi{10.1371/journal.pone.0231866}.
}
\seealso{
Other API functions: 
\code{\link{gb_get}()},
\code{\link{gb_get_cgaz}()}
}
\concept{API functions}
