% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb-get-meta.R
\name{gb_get_meta}
\alias{gb_get_meta}
\title{Get metadata of individual country files from geoBoundaries}
\source{
geoboundaries API Service \url{https://www.geoboundaries.org/api.html}.
}
\usage{
gb_get_meta(
  country = "ALL",
  adm_lvl = c("ALL", "ADM0", "ADM1", "ADM2", "ADM3", "ADM4", "ADM5"),
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative")
)
}
\arguments{
\item{country}{A character vector of country codes. It could be either
\code{"ALL"} (that would return the data for all countries), a vector of country
names or ISO3 country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{adm_lvl}{Type of boundary Accepted values are \code{"ALL"} (all
available boundaries) or the ADM level (\code{"ADM0"} is the country boundary,
\code{"ADM1"} is the first level of sub national boundaries, \code{"ADM2"} is the
second level and so on.}

\item{release_type}{One of \code{"gbOpen"}, \code{"gbHumanitarian"},
\code{"gbAuthoritative"}. Source of the spatial data. See \strong{Details}.}
}
\value{
A tibble.
}
\description{
This function returns metadadata of the
\href{https://www.geoboundaries.org/api.html}{geoBoundaries API}.
}
\details{
Equivalent to \code{gb_get(..., metadata = TRUE)}. See \strong{Details} in \code{\link[=gb_get]{gb_get()}}.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
# Get metadata of ADM4 levels

library(dplyr)

gb_get_meta(adm_lvl = "ADM4") \%>\%
  glimpse()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=gb_get]{gb_get()}}
}
\concept{metadata functions}
