% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb_cache.R
\name{gb_set_cache_dir}
\alias{gb_set_cache_dir}
\title{Set your \pkg{geobounds} cache dir}
\usage{
gb_set_cache_dir(cache_dir, overwrite = FALSE, install = FALSE, quiet = FALSE)
}
\arguments{
\item{cache_dir}{A path to a cache directory. On missing value the function
would store the cached files on a temporary dir (See \code{\link[base:tempfile]{base::tempdir()}}).}

\item{overwrite}{Logical. If this is set to \code{TRUE}, it will overwrite an
existing \code{GEOBOUNDS_CACHE_DIR} that you already have in local machine.}

\item{install}{Logical. If \code{TRUE}, will install the key in your local
machine for use in future sessions.  Defaults to \code{FALSE.} If \code{cache_dir}
is \code{FALSE} this parameter is set to \code{FALSE} automatically.}

\item{quiet}{Logical, on \code{FALSE} it displays information of the call. Useful
for debugging, default is no messages \code{quiet = TRUE}.}
}
\value{
An (invisible) character with the path to your \code{cache_dir}.
}
\description{
This function will store your \code{cache_dir} path on your local machine and
would load it for future sessions. Type
\code{Sys.getenv("GEOBOUNDS_CACHE_DIR")} to find your cached path.

Alternatively, you can store the \code{cache_dir} manually with the following
options:
\itemize{
\item Run \code{Sys.setenv(GEOBOUNDS_CACHE_DIR = "cache_dir")}. You would need to
run this command on each session (Similar to \code{install = FALSE}).
\item Write this line on your .Renviron file:
\code{GEOBOUNDS_CACHE_DIR = "value_for_cache_dir"} (same behavior than
\code{install = TRUE}). This would store your \code{cache_dir} permanently.
}
}
\examples{

# Don't run this! It would modify your current state
\dontrun{
gb_set_cache_dir(quiet = FALSE)
}

Sys.getenv("GEOBOUNDS_CACHE_DIR")
}
\seealso{
\code{\link[tools:userdir]{tools::R_user_dir()}}

Other cache utilities: 
\code{\link{gb_clear_cache}()},
\code{\link{gb_detect_cache_dir}()}
}
\concept{cache utilities}
