% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobounds-package.R
\docType{package}
\name{geobounds-package}
\alias{geobounds}
\alias{geobounds-package}
\title{geobounds: Download Map Data from 'geoBoundaries'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to download data from 'geoBoundaries' \url{https://www.geoboundaries.org/}. Several administration levels available. See Runfola, D. et al. (2020) geoBoundaries: A global database of political administrative boundaries. PLOS ONE 15(4): e0231866. \doi{10.1371/journal.pone.0231866}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/geobounds/}
  \item \url{https://github.com/dieghernan/geobounds}
  \item Report bugs at \url{https://github.com/dieghernan/geobounds/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item geoBoundaries [copyright holder]
}

}
\keyword{internal}
