% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-gb.R
\name{get_gb}
\alias{get_gb}
\title{Get individual country files from geoBoundaries}
\source{
geoboundaries API Service \url{https://www.geoboundaries.org/api.html}.
}
\usage{
get_gb(
  country,
  adm_lvl = c("ADM0", "ADM1", "ADM2", "ADM3", "ADM4", "ALL"),
  simplified = FALSE,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  metadata = FALSE,
  quiet = TRUE,
  overwrite = FALSE,
  cache_dir = NULL
)
}
\arguments{
\item{country}{A character vector of country codes. It could be either
\code{"ALL"} (that would return the data for all countries), a vector of country
names or ISO3 country codes. See also \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{adm_lvl}{Type of boundary Accepted values are \code{"ALL"} (all
available boundaries) or the ADM level (\code{"ADM0"} is the country boundary,
\code{"ADM1"} is the first level of sub national boundaries, \code{"ADM2"} is the
second level and so on.}

\item{simplified}{Logical. Return the simplified boundary or not.}

\item{release_type}{One of \code{"gbOpen"}, \code{"gbHumanitarian"},
\code{"gbAuthoritative"}. Source of the spatial data. See \strong{Details}.}

\item{metadata}{Should the result be the metadata of the boundary?}

\item{quiet}{Logical, on \code{FALSE} it displays information of the call. Useful
for debugging, default is no messages \code{quiet = TRUE}.}

\item{overwrite}{A logical whether to update cache. Default is \code{FALSE}.
When set to \code{TRUE} it would force a fresh download of the source
\code{.geojson} file.}

\item{cache_dir}{A path to a cache directory. If not set (the default
\code{NULL}), the data would be stored in the default cache directory (see
\code{\link[=gb_set_cache_dir]{gb_set_cache_dir()}}). If no cache directory has been set, files would be
stored in the temporary directory (see \code{base::tempdir()}).}
}
\value{
\itemize{
\item With \code{metadata = FALSE}: A \code{\link[sf:sf]{sf}} object.
\item With \code{metadata = TRUE}: A tibble.
}
}
\description{
\href{https://www.geoboundaries.org/index.html#usage}{Attribution} is required
for all uses of this dataset.

This function returns data of individual countries "as they would represent
themselves", with no special identification of disputed areas.

If you would prefer data that explicitly includes disputed areas, please use
\code{\link[=get_gb_cgaz]{get_gb_cgaz()}}.
}
\details{
For most users, we suggest using \code{"gbOpen"}, as it is CC-BY 4.0 compliant,
and can be used for most purposes so long as attribution is provided.
\itemize{
\item \code{"gbHumanitarian"} files are mirrored from
\href{https://www.unocha.org/}{UN OCHA}, but may have less open licensure.
\item \code{"gbAuthoritative"} files are mirrored from
\href{https://salb.un.org/en}{UN SALB}, and cannot  be used for commercial
purposes, but are verified through in-country processes.
}
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}

\donttest{
sri_lanka <- get_gb(
  "Sri Lanka",
  adm_lvl = "ADM3",
  simplified = TRUE
)

sri_lanka

library(ggplot2)
ggplot(sri_lanka) +
  geom_sf() +
  labs(caption = "Source: www.geoboundaries.org")
}

# Metadata
library(dplyr)
get_gb(
  "Sri Lanka",
  adm_lvl = "ADM3",
  metadata = TRUE
) \%>\%
  glimpse()
\dontshow{\}) # examplesIf}
}
\references{
Runfola, D. et al. (2020) geoBoundaries: A global database of political
administrative boundaries. \emph{PLOS ONE} 15(4): e0231866.
\doi{10.1371/journal.pone.0231866}.
}
\seealso{
Other API functions: 
\code{\link{get_gb_adm}},
\code{\link{get_gb_cgaz}()}
}
\concept{API functions}
