% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gb_world.R
\name{get_gb_world}
\alias{get_gb_world}
\title{Get global composites data (CGAZ) from geoBoundaries}
\source{
geoboundaries API Service \url{https://www.geoboundaries.org/api.html}.
}
\usage{
get_gb_world(
  country = "ALL",
  adm_lvl = c("ADM0", "ADM1", "ADM2"),
  quiet = TRUE,
  overwrite = FALSE,
  cache_dir = NULL
)
}
\arguments{
\item{country}{A character vector of country codes. It could be either
\code{"ALL"} (that would return the data for all countries), a vector of country
names or a vector of ISO3 country codes. Mixed types (as
\code{c("Italy","ES","FRA")}) would not work. See also
\code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{adm_lvl}{Type of boundary Accepted values are the ADM level
(\code{"ADM0"} is the country boundary, \code{"ADM1"} is the first level of sub
national boundaries, \code{"ADM2"} is the second level).}

\item{quiet}{Logical, on \code{FALSE} it displays information of the call. Useful
for debugging, default is no messages \code{quiet = TRUE}.}

\item{overwrite}{A logical whether to update cache. Default is \code{FALSE}.
When set to \code{TRUE} it would force a fresh download of the source
\code{.gpkg} file.}

\item{cache_dir}{A path to a cache directory. If not set (the default
\code{NULL}), the data would be stored in the default cache directory (see
\code{\link[=gb_set_cache_dir]{gb_set_cache_dir()}}). If no cache directory has been set, files would be
stored in the temporary directory (see \code{base::tempdir()}).}
}
\value{
A \code{\link[sf:sf]{sf}} object.
}
\description{
\href{https://www.geoboundaries.org/index.html#usage}{Attribution} is required
for all uses of this dataset.

This function returns a global composite of the required administration
level, clipped to international boundaries, with gaps filled between
borders.
}
\details{
Comprehensive Global Administrative Zones (CGAZ) is a set of global
composites for administrative boundaries. There are two important
distinctions between our global product and individual country downloads.
\itemize{
\item Extensive simplification is performed to ensure that file sizes are
small enough to be used in most traditional desktop software.
\item Disputed areas are removed and replaced with polygons following US
Department of State definitions.
}
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}

# This download may take some time
\dontrun{
world <- get_gb_world()

library(ggplot2)

ggplot(world) +
  geom_sf() +
  coord_sf(expand = FALSE) +
  labs(caption = "Source: www.geoboundaries.org")
}
\dontshow{\}) # examplesIf}
}
\references{
Runfola, D. et al. (2020) geoBoundaries: A global database of political
administrative boundaries. \emph{PLOS ONE} 15(4): e0231866.
\doi{10.1371/journal.pone.0231866}.
}
\seealso{
Other API functions: 
\code{\link{get_gb}()},
\code{\link{get_gb_adm}}
}
\concept{API functions}
