% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geobn.R
\name{get_geobn}
\alias{get_geobn}
\title{Get geoBoundaries data}
\usage{
get_geobn(
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  country,
  boundary_type = c("ADM0", "ADM1", "ADM2", "ADM3", "ADM4", "ALL"),
  simplified = FALSE,
  metadata = FALSE,
  verbose = FALSE,
  update_cache = FALSE,
  cache_dir = NULL
)
}
\arguments{
\item{release_type}{One of \code{"gbOpen"}, \code{"gbHumanitarian"},
"gbAuthoritative"`. Source of the spatial data. See \strong{Details}.}

\item{country}{A character vector of country codes. It could be either
\code{"ALL"} (that would return the data for all countries), a vector of country
names or a vector of ISO3 country codes. Mixed types (as
\code{c("Italy","ES","FRA")}) would not work. See also
\code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{boundary_type}{Type of boundary Accepted values are \code{"ALL"} (all
available boundaries) or the ADM level (\code{"ADM0"} is the country boundary,
\code{"ADM1"} is the first level of sub national boundaries, \code{"ADM2"} is the
second level and so on.}

\item{simplified}{Logical. Return the simplified boundary or not.}

\item{metadata}{Should the result be the metadata of the boundary?
See \strong{Details}.}

\item{verbose}{Logical, displays information. Useful for debugging,
default is \code{FALSE}.}

\item{update_cache}{A logical whether to update cache. Default is \code{FALSE}.
When set to \code{TRUE} it would force a fresh download of the source
\code{.geojson} file.}

\item{cache_dir}{A path to a cache directory. See \code{\link[=geobn_set_cache_dir]{geobn_set_cache_dir()}}.}
}
\value{
When \code{metadata = FALSE}: A \code{\link[sf:sf]{sf}} object, otherwise a tibble.
}
\description{
Return an spatial object using \CRANpkg{sf}.
}
\details{
TODO....
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}

# Metadata
library(dplyr)
get_geobn(country = "Spain", boundary_type = "ADM0", metadata = TRUE) \%>\%
  glimpse()
\dontshow{\}) # examplesIf}
}
\references{
geoboundaries API Service \url{https://www.geoboundaries.org/api.html}.
}
