% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geobn_adm.R
\name{get_geobn_adm}
\alias{get_geobn_adm}
\alias{get_geobn_adm0}
\alias{get_geobn_adm1}
\alias{get_geobn_adm2}
\alias{get_geobn_adm3}
\alias{get_geobn_adm4}
\title{Get country files from geoBoundaries for a given administration level}
\source{
geoboundaries API Service \url{https://www.geoboundaries.org/api.html}.
}
\usage{
get_geobn_adm0(
  country,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  simplified = FALSE,
  metadata = FALSE,
  verbose = FALSE,
  update_cache = FALSE,
  cache_dir = NULL
)

get_geobn_adm1(
  country,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  simplified = FALSE,
  metadata = FALSE,
  verbose = FALSE,
  update_cache = FALSE,
  cache_dir = NULL
)

get_geobn_adm2(
  country,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  simplified = FALSE,
  metadata = FALSE,
  verbose = FALSE,
  update_cache = FALSE,
  cache_dir = NULL
)

get_geobn_adm3(
  country,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  simplified = FALSE,
  metadata = FALSE,
  verbose = FALSE,
  update_cache = FALSE,
  cache_dir = NULL
)

get_geobn_adm4(
  country,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  simplified = FALSE,
  metadata = FALSE,
  verbose = FALSE,
  update_cache = FALSE,
  cache_dir = NULL
)
}
\arguments{
\item{country}{A character vector of country codes. It could be either
\code{"ALL"} (that would return the data for all countries), a vector of country
names or a vector of ISO3 country codes. Mixed types (as
\code{c("Italy","ES","FRA")}) would not work. See also
\code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{release_type}{One of \code{"gbOpen"}, \code{"gbHumanitarian"},
"gbAuthoritative"`. Source of the spatial data. See \strong{Details}.}

\item{simplified}{Logical. Return the simplified boundary or not.}

\item{metadata}{Should the result be the metadata of the boundary?}

\item{verbose}{Logical, displays information. Useful for debugging,
default is \code{FALSE}.}

\item{update_cache}{A logical whether to update cache. Default is \code{FALSE}.
When set to \code{TRUE} it would force a fresh download of the source
\code{.geojson} file.}

\item{cache_dir}{A path to a cache directory. See \code{\link[=geobn_set_cache_dir]{geobn_set_cache_dir()}}.}
}
\value{
\itemize{
\item With \code{metadata = FALSE}: A \code{\link[sf:sf]{sf}} object.
\item With \code{metadata = TRUE}: A tibble.
}
}
\description{
\href{https://www.geoboundaries.org/index.html#usage}{Attribution} is required
for all uses of this dataset.

These functions are wrappers of \code{\link[=get_geobn]{get_geobn()}} for extracting any
given administration level:
\itemize{
\item \code{get_geobn_adm0()} returns the country boundary.
\item \code{get_geobn_adm1()} returns first-level administration
boundaries (e.g. States in the United States).
\item \code{get_geobn_adm2()} returns second-level administration
boundaries (e.g. Counties in the United States).
\item \code{get_geobn_adm3()} returns third-level administration
boundaries (e.g. towns or cities in some countries).
\item \code{get_geobn_adm3()} returns fourth-level administration
boundaries.
}

Note that not all countries have the same number of levels.
}
\details{
See \strong{Details} in \code{\link[=get_geobn]{get_geobn()}}.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}

\donttest{
lev2 <- get_geobn_adm2(
  c("Italia", "Suiza", "Austria"),
  simplified = TRUE
)


library(ggplot2)

ggplot(lev2) +
  geom_sf(aes(fill = shapeGroup)) +
  labs(
    title = "Second-level administration",
    subtitle = "Selected countries",
    caption = "Source: www.geoboundaries.org"
  )
}
\dontshow{\}) # examplesIf}
}
\references{
Runfola, D. et al. (2020) geoBoundaries: A global database of political
administrative boundaries. \emph{PLoS ONE} \strong{15}(4): e0231866.
\doi{10.1371/journal.pone.0231866}.
}
\seealso{
Other API functions: 
\code{\link{get_geobn}()},
\code{\link{get_geobn_world}()}
}
\concept{API functions}
