% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geobn_meta.R
\name{get_geobn_meta}
\alias{get_geobn_meta}
\title{Get metadata of individual country files from geoBoundaries}
\source{
geoboundaries API Service \url{https://www.geoboundaries.org/api.html}.
}
\usage{
get_geobn_meta(
  country = "ALL",
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  boundary_type = c("ALL", "ADM0", "ADM1", "ADM2", "ADM3", "ADM4")
)
}
\arguments{
\item{country}{A character vector of country codes. It could be either
\code{"ALL"} (that would return the data for all countries), a vector of country
names or a vector of ISO3 country codes. Mixed types (as
\code{c("Italy","ES","FRA")}) would not work. See also
\code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{release_type}{One of \code{"gbOpen"}, \code{"gbHumanitarian"},
"gbAuthoritative"`. Source of the spatial data. See \strong{Details}.}

\item{boundary_type}{Type of boundary Accepted values are \code{"ALL"} (all
available boundaries) or the ADM level (\code{"ADM0"} is the country boundary,
\code{"ADM1"} is the first level of sub national boundaries, \code{"ADM2"} is the
second level and so on.}
}
\value{
A tibble.
}
\description{
This function returns metadadata of the
\href{https://www.geoboundaries.org/api.html}{geoBoundaries API}.
}
\details{
Equivalent to \code{get_geobn(..., metadata = TRUE)}.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
# Get ADM4 levels

library(dplyr)

get_geobn_meta(boundary_type = "ADM4") \%>\%
  glimpse()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=get_geobn]{get_geobn()}}
}
\concept{metadata functions}
