% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geobn_world.R
\name{get_geobn_world}
\alias{get_geobn_world}
\title{Get global composites data from geoBoundaries}
\usage{
get_geobn_world(
  boundary_type = c("ADM0", "ADM1", "ADM2"),
  verbose = FALSE,
  update_cache = FALSE,
  cache_dir = NULL
)
}
\arguments{
\item{boundary_type}{Type of boundary Accepted values are the ADM level
(\code{"ADM0"} is the country boundary, \code{"ADM1"} is the first level of sub
national boundaries, \code{"ADM2"} is the second level).}

\item{verbose}{Logical, displays information. Useful for debugging,
default is \code{FALSE}.}

\item{update_cache}{A logical whether to update cache. Default is \code{FALSE}.
When set to \code{TRUE} it would force a fresh download of the source
\code{.gpkg} file.}

\item{cache_dir}{A path to a cache directory. See \code{\link[=geobn_set_cache_dir]{geobn_set_cache_dir()}}.}
}
\value{
A \code{\link[sf:sf]{sf}} object.
}
\description{
\href{https://www.geoboundaries.org/index.html#usage}{Attribution} is required
for all uses of this dataset.

This function returns a global composite of the required administration level,
clipped to international boundaries, with gaps filled between borders.
}
\details{
Comprehensive Global Administrative Zones (CGAZ) is a set of global
composites for administrative boundaries. There are two important
distinctions between our global product and individual country downloads.
\itemize{
\item Extensive simplification is performed to ensure that file sizes are
small enough to be used in most traditional desktop software.
\item Disputed areas are removed and replaced with polygons following US
Department of State definitions.
}
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}

# This download may take some time
\dontrun{
world <- get_geobn_world()

library(ggplot2)

ggplot(world) +
  geom_sf() +
  coord_sf(expand = FALSE) +
  labs(caption = "Source: www.geoboundaries.org")
}
\dontshow{\}) # examplesIf}
}
\references{
geoboundaries API Service \url{https://www.geoboundaries.org/api.html}.

Runfola, D. et al. (2020) geoBoundaries: A global database of political
administrative boundaries. \emph{PLoS ONE} \strong{15}(4): e0231866.
\doi{10.1371/journal.pone.0231866}.
}
\seealso{
Other API functions: 
\code{\link{get_geobn}()},
\code{\link{get_geobn_adm}}
}
\concept{API functions}
