% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygeoboundaries-package.R
\docType{package}
\name{tidygeoboundaries-package}
\alias{tidygeoboundaries}
\alias{tidygeoboundaries-package}
\title{tidygeoboundaries: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/tidygeoboundaries/}
  \item \url{https://github.com/dieghernan/tidygeoboundaries}
  \item Report bugs at \url{https://github.com/dieghernan/tidygeoboundaries/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

}
\keyword{internal}
