% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb_cache_clear.R
\name{gb_clear_cache}
\alias{gb_clear_cache}
\title{Clear your \pkg{geobounds} cache dir}
\usage{
gb_clear_cache(config = FALSE, cached_data = TRUE, quiet = TRUE)
}
\arguments{
\item{config}{Logical. If \code{TRUE}, will delete the configuration folder of
\pkg{geobounds}.}

\item{cached_data}{Logical. If \code{TRUE}, it will delete your \code{cache_dir} and
all its content.}

\item{quiet}{Logical, on \code{FALSE} it displays information of the call. Useful
for debugging, default is no messages \code{quiet = TRUE}.}
}
\value{
Invisible. This function is called for its side effects.
}
\description{
\strong{Use this function with caution}. This function would clear your cached
data and configuration, specifically:
\itemize{
\item Deletes the \pkg{geobounds} config directory
(\code{tools::R_user_dir("geobounds", "config")}).
\item Deletes the \code{cache_dir} directory.
\item Deletes the values on stored on \code{Sys.getenv("GEOBOUNDS_CACHE_DIR")}.
}
}
\details{
This is an overkill function that is intended to reset your status
as it you would never have installed and/or used \pkg{geobounds}.
}
\examples{

# Don't run this! It would modify your current state
\dontrun{
gb_clear_cache(quiet = FALSE)
}

}
\seealso{
Other cache utilities: 
\code{\link{gb_detect_cache_dir}()},
\code{\link{gb_set_cache_dir}()}
}
\concept{cache utilities}
