% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geobn.R
\name{get_geobn}
\alias{get_geobn}
\title{Get individual country files from geoBoundaries}
\source{
geoboundaries API Service \url{https://www.geoboundaries.org/api.html}.
}
\usage{
get_geobn(
  country,
  release_type = c("gbOpen", "gbHumanitarian", "gbAuthoritative"),
  boundary_type = c("ADM0", "ADM1", "ADM2", "ADM3", "ADM4", "ALL"),
  simplified = FALSE,
  metadata = FALSE,
  verbose = FALSE,
  update_cache = FALSE,
  cache_dir = NULL
)
}
\arguments{
\item{country}{A character vector of country codes. It could be either
\code{"ALL"} (that would return the data for all countries), a vector of country
names or a vector of ISO3 country codes. Mixed types (as
\code{c("Italy","ES","FRA")}) would not work. See also
\code{\link[countrycode:countrycode]{countrycode::countrycode()}}.}

\item{release_type}{One of \code{"gbOpen"}, \code{"gbHumanitarian"},
"gbAuthoritative"`. Source of the spatial data. See \strong{Details}.}

\item{boundary_type}{Type of boundary Accepted values are \code{"ALL"} (all
available boundaries) or the ADM level (\code{"ADM0"} is the country boundary,
\code{"ADM1"} is the first level of sub national boundaries, \code{"ADM2"} is the
second level and so on.}

\item{simplified}{Logical. Return the simplified boundary or not.}

\item{metadata}{Should the result be the metadata of the boundary?}

\item{verbose}{Logical, displays information. Useful for debugging,
default is \code{FALSE}.}

\item{update_cache}{A logical whether to update cache. Default is \code{FALSE}.
When set to \code{TRUE} it would force a fresh download of the source
\code{.geojson} file.}

\item{cache_dir}{A path to a cache directory. See \code{\link[=geobn_set_cache_dir]{geobn_set_cache_dir()}}.}
}
\value{
\itemize{
\item With \code{metadata = FALSE}: A \code{\link[sf:sf]{sf}} object.
\item With \code{metadata = TRUE}: A tibble.
}
}
\description{
\href{https://www.geoboundaries.org/index.html#usage}{Attribution} is required
for all uses of this dataset.

This function returns data of individual countries "as they would represent
themselves", with no special identification of disputed areas.
}
\details{
For most users, we suggest using \code{"gbOpen"}, as it is CC-BY 4.0 compliant,
and can be used for most purposes so long as attribution is provided.
\code{"gbHumanitarian"} files are mirrored from
\href{https://www.unocha.org/}{UN OCHA}, but may have less open licensure.
\code{"gbAuthoritative"} files are mirrored from
\href{https://salb.un.org/en}{UN SALB}, and cannot  be used for commercial
purposes, but are verified through in-country processes.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}

\donttest{
sri_lanka <- get_geobn(
  "Sri Lanka",
  boundary_type = "ADM3",
  simplified = TRUE
)

sri_lanka

library(ggplot2)
ggplot(sri_lanka) +
  geom_sf() +
  labs(caption = "Source: www.geoboundaries.org")
}

# Metadata
library(dplyr)
get_geobn(
  "Sri Lanka",
  boundary_type = "ADM3",
  metadata = TRUE
) \%>\%
  glimpse()
\dontshow{\}) # examplesIf}
}
\references{
Runfola, D. et al. (2020) geoBoundaries: A global database of political
administrative boundaries. \emph{PLoS ONE} \strong{15}(4): e0231866.
\doi{10.1371/journal.pone.0231866}.
}
\seealso{
Other API functions: 
\code{\link{get_geobn_adm}},
\code{\link{get_geobn_world}()}
}
\concept{API functions}
